+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- The default value for `alphaMode` on `StandardMaterial` was changed to `blend` to make it easier to render transparent objects without having to change `alphaMode`. To get the previous behavior, set `alphaMode` to `opaque`.
+
+## [1.5.1] - 2022-07-26
+### Fixed
+- Removed peer dependencies from `package.json` which caused issues with multiple PixiJS versions when using bundlers.
+
+## [1.5.0] - 2022-07-23
 ### Added
 - Added `Mesh3D.createSphere` function for creating a sphere mesh.
 - Added support for vertex colors.